requirements.txt
Django==5.0.6
django-extensions<=3.2.3
djangorestframework==3.15.1
requests==2.32.3

html2text==2024.2.26
django-environ
requests-oauthlib==2.0.0
stripe==10.5.0